Pull out expression modifying tree visitor
